recommender_bq/requirements.txt
google-cloud-recommender
pandas
google-cloud-bigquery
google-cloud-secret-manager
db-dtypes